requirements.txt

email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gensim>=4.3.3
gunicorn>=23.0.0
joblib>=1.4.2
numpy>=1.26.4
pandas>=2.2.3
psycopg2-binary>=2.9.10
scikit-learn>=1.6.1
tldextract>=5.1.3
trafilatura>=2.0.0
validators>=0.34.0
flask-wtf>=1.2.2
sqlalchemy>=2.0.38
werkzeug>=3.1.3
flask-migrate
python-dotenv
